re-organize directory and file structure
