--- src/components/Coffees/Coffee.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-function App() {
-  return (
-    <div>
-      <h2>A Coffee</h2>
-    </div>
-  );
-}
-
-export default App;
