--- src/components/Methods/Method.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-function App() {
-  return (
-    <div>
-      <h2>A Method</h2>
-    </div>
-  );
-}
-
-export default App;
