--- src/components/Recipes/Recipe.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-function App() {
-  return (
-    <div>
-      <h2>A Recipe</h2>
-    </div>
-  );
-}
-
-export default App;
