+++ src/views/Methods.js
